realized how silly fallback alloc is
it was O(n) due to memmove() whenever using fallback. so binary
searching was kind of pointless.

also, the reason to have a fallback_allocations array was to know the
size... but realloc is typically used when the "client code" already
knows what the old/new size is, so lets leave that as an exercise for
THAT code (maybe I should get rid of realloc altogether?)

instead I now have a slab_can_alloc() which also knows about degenerate
cases where all slabs are full